Split the example app code into files
