+++ auth0_flutter/example/lib/constants.dart
@@ -1,0 +1,1 @@
+const double padding = 24.0;
